requirements.txt
Django==4.2.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
openai==0.27.6
python-dotenv==1.0.0